requirements.txt
Flask == 1.1.2
SQLAlchemy == 1.3.23
PyMySQL == 1.0.2
python-dotenv == 0.15.0
Flask-SQLAlchemy == 2.4.4
bcrypt == 3.2.0
Flask-Login == 0.5.0
cryptography == 3.4.5
email-validator == 1.1.2
alembic==1.6.5
click==8.0.1
Flask-Migrate==3.0.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
python-dateutil==2.8.2
python-editor==1.0.4
six==1.16.0
Werkzeug==2.0.1